# swagger/swagger-output.yaml
openapi: 3.0.0
info:
  title: School Management API
  version: 1.0.0
  description: API for school admin storage and management
servers:
  - url: https://cse-341-group-project-mfvd.onrender.com/api
    description: Production server
  - url: http://localhost:5500/api
    description: Local development server
components:
  securitySchemes:
    googleOAuth:
      type: oauth2
      flows:
        authorizationCode:
          authorizationUrl: https://accounts.google.com/o/oauth2/v2/auth
          tokenUrl: https://oauth2.googleapis.com/token
          scopes:
            openid: OpenID access
            profile: View your profile info
            email: View your email address
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    user:
      type: object
      properties:
        _id:
          type: string
          description: Autogenerated ID of the user
          readOnly: true
        name:
          type: string
          description: Full name of the user
          example: Jane Doe
        email:
          type: string
          format: email
          description: Email address
          example: janedoe@school.edu
        password:
          type: string
          description: Hashed password
          example: $2a$10$abc...
          writeOnly: true
        role:
          type: string
          enum:
            - student
            - staff
            - admin
          description: User role
          example: student
        refId:
          type: string
          description: Reference ID to role-specific document
          example: 60f6e9f2b7a1c142d8ef8c00
        isActive:
          type: boolean
          description: User active status
          example: true
        createdAt:
          type: string
          format: date-time
          readOnly: true
        updatedAt:
          type: string
          format: date-time
          readOnly: true
      required:
        - name
        - email
        - password
        - role
        - refId
...
    course:
      type: object
      properties:
        _id:
          type: string
          readOnly: true
        title:
          type: string
          example: Algebra 101
        description:
          type: string
          example: Introductory algebra course
        department:
          type: string
          example: 60f6e9f2b7a1c142d8ef8c11
        instructor:
          type: string
          example: 60f6e9f2b7a1c142d8ef8c12
        createdAt:
          type: string
          format: date-time
          readOnly: true
        updatedAt:
          type: string
          format: date-time
          readOnly: true
      required:
        - title
        - department
        - instructor
    enrollment:
      type: object
      properties:
        _id:
          type: string
          readOnly: true
        student:
          type: string
          example: 60f6e9f2b7a1c142d8ef8c10
        course:
          type: string
          example: 60f6e9f2b7a1c142d8ef8c11
        enrolledAt:
          type: string
          format: date-time
          example: '2025-01-10T12:00:00Z'
        grade:
          type: string
          enum:
            - A
            - B
            - C
            - D
            - E
            - F
            - I
            - W
            - null
          example: A
        status:
          type: string
          enum:
            - enrolled
            - completed
            - withdrawn
          example: enrolled
        createdAt:
          type: string
          format: date-time
          readOnly: true
        updatedAt:
          type: string
          format: date-time
          readOnly: true
      required:
        - student
        - course
security:
  - bearerAuth: []
paths:
...
  /course/{id}:
    get:
      summary: Get a course by ID
      tags:
        - Course
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          description: The course ID
      responses:
        '200':
          description: Course found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/course'
        '404':
          description: Course not found
    put:
      summary: Update a course by ID
      tags:
        - Course
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          description: The course ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/course'
      responses:
        '200':
          description: Course updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/course'
        '400':
          description: Invalid input
        '404':
          description: Course not found
    delete:
      summary: Delete a course by ID
      tags:
        - Course
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          description: The course ID
      responses:
        '204':
          description: Course deleted successfully
        '404':
          description: Course not found
...
  /enrollment:
    post:
      summary: Enroll a student in a course
      tags:
        - Enrollment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/enrollment'
      responses:
        '201':
          description: Student enrolled successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/enrollment'
        '400':
          description: Invalid input
        '404':
          description: Student or course not found
    get:
      summary: Get all enrollments
      tags:
        - Enrollment
      responses:
        '200':
          description: List of all enrollments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/enrollment'
  /enrollment/student/{studentId}:
    get:
      summary: Get enrollments for a specific student
      tags:
        - Enrollment
      parameters:
        - in: path
          name: studentId
          required: true
          schema:
            type: string
          description: The student ID
      responses:
        '200':
          description: List of student's enrollments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/enrollment'
        '404':
          description: Student not found
  /enrollment/course/{courseId}:
    get:
      summary: Get enrollments for a specific course
      tags:
        - Enrollment
      parameters:
        - in: path
          name: courseId
          required: true
          schema:
            type: string
          description: The course ID
      responses:
        '200':
          description: List of course enrollments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/enrollment'
        '404':
          description: Course not found
  /enrollment/{id}:
    put:
      summary: Update enrollment (grade or status)
      tags:
        - Enrollment
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          description: The enrollment ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/enrollment'
      responses:
        '200':
          description: Enrollment updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/enrollment'
        '400':
          description: Invalid input
        '404':
          description: Enrollment not found
  /enrollment/{id}/withdraw:
    patch:
      summary: Withdraw a student from a course
      tags:
        - Enrollment
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          description: The enrollment ID
      responses:
        '200':
          description: Student withdrawn successfully
        '404':
          description: Enrollment not found
  /user:
    post:
      summary: Create a new user
      tags:
        - User
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/user'
      responses:
        '201':
          description: User created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/user'
        '400':
          description: Invalid input
    get:
      summary: Get all users
      tags:
        - User
      responses:
        '200':
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/user'
  /user/signup:
    post:
      summary: Register a new user and get a JWT token
      tags:
        - User
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - email
                - password
                - role
              properties:
                name:
                  type: string
                  example: Alex Okhitoya
                email:
                  type: string
                  format: email
                  example: alex@school.com
                password:
                  type: string
                  format: password
                  example: secret123
                role:
                  type: string
                  enum:
                    - admin
                    - teacher
                    - student
                  example: student
      responses:
        '201':
          description: User registered successfully and JWT token returned
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  token:
                    type: string
                    example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
                  user:
                    type: object
                    properties:
                      id:
                        type: string
                        example: 60e7a2a5b5e4c24d88e2f776
                      name:
                        type: string
                        example: Alex Okhitoya
                      email:
                        type: string
                        example: alex@school.com
                      role:
                        type: string
                        example: student
        '400':
          description: User already exists or validation failed
        '500':
          description: Server error during signup
  /user/{email}:
    get:
      summary: Get a user by email
      tags:
        - User
      parameters:
        - in: path
          name: email
          required: true
          schema:
            type: string
          description: The user email
      responses:
        '200':
          description: User found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/user'
        '404':
          description: User not found
  /user/{id}:
    get:
      summary: Get a user by ID
      tags:
        - User
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          description: The user ID
      responses:
        '200':
          description: User found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/user'
        '404':
          description: User not found
    put:
      summary: Update a user by ID
      tags:
        - User
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          description: The user ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/user'
      responses:
        '200':
          description: User updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/user'
        '400':
          description: Invalid input
        '404':
          description: User not found
    delete:
      summary: Delete a user member by ID
      tags:
        - User
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          description: The user ID
      responses:
        '200':
          description: User member deleted
        '404':
          description: User not found
...